pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["dbp"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
markdown = "^3.4.1"
drf-yasg = {extras = ["validation"], version = "^1.21.5"}
djangorestframework-simplejwt = "^5.2.2"
coverage = {version = "6.5.0", extras = ["toml"]}
coveralls = "^3.3.1"
rollbar = "^0.16.3"
python-dotenv = "^1.0.0"
newrelic = "^8.7.0"
django-storages = "^1.13.2"
boto3 = "^1.26.90"
pillow = "^9.4.0"
drf-extensions = "^0.7.1"
celery = "^5.2.7"
redis = "^4.5.2"
flower = "^1.2.0"


[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
freezegun = "^1.2.2"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
relative_files = true